requirements.txt
scipy>=0.18.1
numpy>=1.11.2
pytest>=3.0
pandas>=1.0.0
matplotlib>=3.1.3
biopython>=1.76
